MAJOR UPGRADE to TComponent
This adds global 'fx' events, behaviors, class-wide behaviors, intra-object/behavior 'dy' events, and tweaks the raiseEvent to be more advanced (yet backward compatible).  A large amount of documentation is in here too.  Slight changes to the TApplication object is  made to provide events with priorities.  TComponent has a massive amount of test cases as well…  There are now 512 assertions.  Test cases have been added for testing the GetJS and SetJS and ensure functionality of these new getters and setters.

The Behaviors are all tested and fully integrated into the TComponent Object.  Behaviors are more transparently and widely integrated into TComponent.

Class Behaviors rely on the global events to add behaviors to instanced objects.  Class behaviors also allow for behaviors to be added during the instancing of new objects of the class.

Dynamic events enable the object to send events directly to behaviors as intra-object events.  Dynamic events can be chained as multiple behaviors may implement the event.  This allows for parameter filtering.

RaiseEvent allows for raising global events, capturing results, and feeding results forward into the next handlers parameters should the need arise.  This is a forward thinking feature to enable new processes to be designed into the code for PRADO version 3.3 and future version 4.0.  This functionality is transparent when unused.

We are hoping this is fully backward compatible with the existing code base.  It was designed for maximum backward compatibility while expanding the functional base exponentially.

This clears the way for a new way of thinking about how objects relate.